requirements.txt
livekit-agents[openai, silero, turn-detector,elevenlabs,groq]>=1.0
livekit-plugins-soniox>=1.2.11
python-dotenv>=1.0
sounddevice>=0.4.6
scipy>=1.11.0
certifi>=2023.7.22
numpy>=1.24.0

# Data handling
pydantic
dataclasses-json

# Optional: for development
pytest
black
flake8